requirements.txt
aws-cdk-lib==2.95.1
aws-cdk.aws-kinesisfirehose-alpha==2.95.1a0
aws-cdk.aws-kinesisfirehose-destinations-alpha==2.95.1a0
constructs>=10.0.0,<11.0.0
